Refactor: extract Utxo to it's own file
